api跑通

+++ src/main/java/com/yss/Test.java
@@ -8,7 +8,7 @@ public class Test {
     public static void main(String[] args) {
         AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(SpringConfig.class);
         SecKillDao secKillDao = (SecKillDao)context.getBean(SecKillDao.class);
-        System.out.println(secKillDao.selectOne(1000));
+        System.out.println(secKillDao.selectOne((long) 1000));
         System.out.println(1);
     }
